requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.1.7
djangorestframework==3.15.2
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.4.0
gunicorn==23.0.0
mysqlclient==2.2.7
PyJWT==2.10.1
cryptography==45.0.4
kickbox==2.0.4